requirements.txt
asgiref==3.5.2
certifi==2022.6.15.2
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==3.2.9
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-cloudinary-storage==0.3.0
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.3
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==21.3
Pillow==9.2.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==1.5
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
